+++ packages/vue/src/primitives/ScrubInput/index.ts
@@ -1,5 +1,10 @@
-export { default as ScrubInputRoot } from '#vue/primitives/ScrubInput/ScrubInputRoot.vue'
-export { default as ScrubInputField } from '#vue/primitives/ScrubInput/ScrubInputField.vue'
-export { default as ScrubInputDisplay } from '#vue/primitives/ScrubInput/ScrubInputDisplay.vue'
-export { useScrubInput } from '#vue/primitives/ScrubInput/context'
-export type { ScrubInputContext } from '#vue/primitives/ScrubInput/context'
+/** @deprecated Use NumberFieldRoot. */
+export { default as ScrubInputRoot } from '#vue/primitives/NumberField/NumberFieldRoot.vue'
+/** @deprecated Use NumberFieldInput. */
+export { default as ScrubInputField } from '#vue/primitives/NumberField/NumberFieldInput.vue'
+/** @deprecated Use NumberFieldValue. */
+export { default as ScrubInputDisplay } from '#vue/primitives/NumberField/NumberFieldValue.vue'
+/** @deprecated Use useNumberField. */
+export { useNumberField as useScrubInput } from '#vue/primitives/NumberField/context'
+/** @deprecated Use NumberFieldContext. */
+export type { NumberFieldContext as ScrubInputContext } from '#vue/primitives/NumberField/types'